requirements.txt
selenium
requests
beautifulsoup4
argparse
Pillow
dotenv
webdriver_manager